fix(lsp): goto definition on variables in `match guard`
