+++ tests/fixtures/lsp/goto/else.nu
@@ -1,0 +1,2 @@
+let foo = 1
+if true { } else { $foo }
