+++ tests/fixtures/lsp/goto/match.nu
@@ -1,0 +1,5 @@
+let foo = 1
+match $foo {
+  _ if $foo == 1 => 1
+  _ => 2
+}
